// types/api.types.ts
export type CategoryT = {
  id: number;
  name: string;
  description: string;
  shortDescription: string;
  thumbnail: string;
  children: CategoryT[];
};
// ...
export type PageSectionItemT = {
  id: number;
  sectionId: number;
  type: number;
  status: number;
  sortOrder: number | null;
  name: string | null;
  description: string | null;
  thumbnail: string | null;
  mediaAlt: string | null;
  mediaType: string | null;
  mediaPath: string | null;
  startDate: string | null;
  endDate: string | null;
  linkUrl: string | null;
  itemGUID: string | null;
  itemId: number;
  itemType: number;
  pageId: number;
  jsonExt: string | null;
  setting: string | null;
  locale: string | null;
  createdOn: string;
};

export type PageSectionT = {
  id: number;
  parentId: number | null;
  title: string | null;
  status: number;
  keyword: string | null;
  pageId: number;
  startDate: string | null;
  endDate: string | null;
  type: number;
  version: number | null;
  setting: string | null;
  description: string | null;
  sortOrder: number | null;
  locale: string | null;
  createdOn: string;
  items: PageSectionItemT[];
};